ml-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
torch>=2.0.0
torchvision>=0.15.0
tensorflow>=2.15.0
opencv-python>=4.8.0
numpy>=1.24.0
pillow>=10.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
aiofiles>=23.2.0
timm>=0.9.0
einops>=0.7.0
scipy>=1.11.0
